Moved fakes creator into a single file
